✨ 채팅 기능 구현 서버 설정 전
